Remove parts and tools from topic schema
Because each procedure has a list of tools and parts that are required
for the procedure, a topic-level tools and parts list is unecessary.  If
the implementor would like a topic-level Bill of Materials or tools
list, they can compile it from the list of procedures.  This also
verifies that the topic level BoM is always up to date.